deployment/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
azure-ai-formrecognizer==3.3.2
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.15.0
azure-mgmt-cognitiveservices==13.5.0
azure-mgmt-core==1.4.0
azure-keyvault==4.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.5
msal==1.27.0
msal-extensions==1.1.0
msrest==0.7.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.13.3
orjson==3.9.15
packaging==23.2
portalocker==2.8.2
pycparser==2.21
pydantic==2.6.3
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.8.0
pymongo==4.6.2
python-docx==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tenacity==8.2.3
tqdm==4.66.2
typing_extensions==4.10.0
ujson==5.9.0
urllib3==2.2.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
uvicorn
gunicorn
fastapi
